another one

+++ spring-demo/src/main/java/ro/tuc/ds2022/controllers/handlers/dto/DeviceMeasurementDto.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class DeviceMeasurementDto {
     private Long device_id;
     private String timestamp;
-    private Long energyCon;
+    private Float energyCon;
 
 }
