+++ spring-demo/src/main/java/ro/tuc/ds2022/controllers/handlers/dto/MeasurementDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class MeasurementDto {
 
-   private Long energyCon;
+   private Float energyCon;
    private String  unityOfM;
 
    private String timestamp;
